+++ modules/main/src/main/java/com/nisovin/shopkeepers/shopobjects/living/LivingEntityShopListener.java
@@ -3,7 +3,6 @@ package com.nisovin.shopkeepers.shopobjects.living;
 import java.util.Collection;
 
 import org.bukkit.Bukkit;
-import org.bukkit.Material;
 import org.bukkit.block.Block;
 import org.bukkit.entity.Entity;
 import org.bukkit.entity.LivingEntity;
@@ -42,8 +41,6 @@ import org.bukkit.event.player.PlayerInteractAtEntityEvent;
 import org.bukkit.event.player.PlayerInteractEntityEvent;
 import org.bukkit.event.vehicle.VehicleEnterEvent;
 import org.bukkit.inventory.EquipmentSlot;
-import org.bukkit.inventory.ItemStack;
-import org.bukkit.inventory.PlayerInventory;
 import org.bukkit.projectiles.ProjectileSource;
 import org.checkerframework.checker.nullness.qual.NonNull;
 import org.checkerframework.checker.nullness.qual.Nullable;
@@ -195,19 +192,6 @@ class LivingEntityShopListener implements Listener {
 			return;
 		}
 
-		// TODO Minecraft bug: https://bugs.mojang.com/browse/MC-141494
-		// Interacting with a villager while holding a written book in the main or off hand results
-		// in weird glitches and tricks the plugin into thinking that the editor or trading UI got
-		// opened even though the book got opened instead. We therefore ignore any interactions with
-		// shopkeeper mobs for now when the interacting player is holding a written book.
-		// TODO This has been fixed in MC 1.16. Remove this check once we only support MC 1.16 and
-		// above.
-		if (hasWrittenBookInHand(player)) {
-			Log.debug("  Ignoring interaction due to holding a written book in main or off hand. "
-					+ "See Minecraft issue MC-141494.");
-			return;
-		}
-
 		// Check the entity interaction result by calling another interact event:
 		if (Settings.checkShopInteractionResult) {
 			if (!InteractionUtils.checkEntityInteract(player, clickedEntity)) {
@@ -218,17 +202,6 @@ class LivingEntityShopListener implements Listener {
 
 		// Handle interaction:
 		shopkeeper.onPlayerInteraction(player);
-	}
-
-	private static boolean hasWrittenBookInHand(Player player) {
-		assert player != null;
-		PlayerInventory inventory = player.getInventory();
-		return isWrittenBook(inventory.getItemInMainHand())
-				|| isWrittenBook(inventory.getItemInOffHand());
-	}
-
-	private static boolean isWrittenBook(@Nullable ItemStack itemStack) {
-		return (itemStack != null) && (itemStack.getType() == Material.WRITTEN_BOOK);
 	}
 
 	// This event gets sometimes called additionally to the PlayerInteractEntityEvent