+++ modules/main/src/main/java/com/nisovin/shopkeepers/shopobjects/living/types/VillagerShop.java
@@ -18,10 +18,6 @@ import com.nisovin.shopkeepers.api.ui.DefaultUITypes;
 import com.nisovin.shopkeepers.config.Settings;
 import com.nisovin.shopkeepers.lang.Messages;
 import com.nisovin.shopkeepers.shopkeeper.AbstractShopkeeper;
-import com.nisovin.shopkeepers.shopkeeper.ShopkeeperData;
-import com.nisovin.shopkeepers.shopkeeper.migration.Migration;
-import com.nisovin.shopkeepers.shopkeeper.migration.MigrationPhase;
-import com.nisovin.shopkeepers.shopkeeper.migration.ShopkeeperDataMigrator;
 import com.nisovin.shopkeepers.shopobjects.ShopObjectData;
 import com.nisovin.shopkeepers.shopobjects.living.LivingShops;
 import com.nisovin.shopkeepers.shopobjects.living.SKLivingShopObjectType;
@@ -41,7 +37,6 @@ import com.nisovin.shopkeepers.util.data.serialization.java.NumberSerializers;
 import com.nisovin.shopkeepers.util.inventory.ItemUtils;
 import com.nisovin.shopkeepers.util.java.EnumUtils;
 import com.nisovin.shopkeepers.util.java.MathUtils;
-import com.nisovin.shopkeepers.util.logging.Log;
 
 public class VillagerShop extends BabyableShop<@NonNull Villager> {
 
@@ -63,52 +58,6 @@ public class VillagerShop extends BabyableShop<@NonNull Villager> {
 			.validator(IntegerValidators.bounded(MIN_VILLAGER_LEVEL, MAX_VILLAGER_LEVEL))
 			.defaultValue(1)
 			.build();
-
-	static {
-		// Register shopkeeper data migrations:
-		ShopkeeperDataMigrator.registerMigration(new Migration(
-				"villager-profession",
-				MigrationPhase.ofShopObjectClass(VillagerShop.class)
-		) {
-			@Override
-			public boolean migrate(
-					ShopkeeperData shopkeeperData,
-					String logPrefix
-			) throws InvalidDataException {
-				ShopObjectData shopObjectData = shopkeeperData.get(AbstractShopkeeper.SHOP_OBJECT_DATA);
-				boolean migrated = false;
-
-				// Migration from 'prof' key: TODO Added with 1.14 update, remove again at some
-				// point.
-				String professionName = shopObjectData.getString("prof");
-				if (professionName != null) {
-					Log.warning(logPrefix
-							+ "Migrated villager profession from key 'prof' to key 'profession'.");
-					shopObjectData.remove("prof");
-					shopObjectData.set(DATA_KEY_PROFESSION, professionName);
-					migrated = true;
-				}
-
-				// MC 1.14 migration:
-				professionName = shopObjectData.getString(DATA_KEY_PROFESSION);
-				if (professionName != null) {
-					Profession newProfession = null;
-					if (professionName.equals("PRIEST")) {
-						newProfession = Profession.CLERIC;
-					} else if (professionName.equals("BLACKSMITH")) {
-						newProfession = Profession.ARMORER;
-					}
-					if (newProfession != null) {
-						Log.warning(logPrefix + "Migrated villager profession from '"
-								+ professionName + "' to '" + newProfession.name() + "'.");
-						shopObjectData.set(PROFESSION, newProfession);
-						migrated = true;
-					}
-				}
-				return migrated;
-			}
-		});
-	}
 
 	private final PropertyValue<@NonNull Profession> professionProperty = new PropertyValue<>(PROFESSION)
 			.onValueChanged(Unsafe.initialized(this)::applyProfession)