+++ modules/main/src/main/java/com/nisovin/shopkeepers/shopobjects/sign/SignType.java
@@ -5,7 +5,6 @@ import java.util.function.Predicate;
 import org.bukkit.Material;
 import org.checkerframework.checker.nullness.qual.Nullable;
 
-import com.nisovin.shopkeepers.compat.MC_1_16;
 import com.nisovin.shopkeepers.util.java.Validate;
 
 public enum SignType {
@@ -18,8 +17,8 @@ public enum SignType {
 	JUNGLE(Material.JUNGLE_SIGN, Material.JUNGLE_WALL_SIGN),
 	ACACIA(Material.ACACIA_SIGN, Material.ACACIA_WALL_SIGN),
 	DARK_OAK(Material.DARK_OAK_SIGN, Material.DARK_OAK_WALL_SIGN),
-	CRIMSON(MC_1_16.CRIMSON_SIGN, MC_1_16.CRIMSON_WALL_SIGN),
-	WARPED(MC_1_16.WARPED_SIGN, MC_1_16.WARPED_WALL_SIGN);
+	CRIMSON(Material.CRIMSON_SIGN, Material.CRIMSON_WALL_SIGN),
+	WARPED(Material.WARPED_SIGN, Material.WARPED_WALL_SIGN);
 
 	public static final Predicate<SignType> IS_SUPPORTED = SignType::isSupported;